pwm: add support for AXI PWM generator
This patch adds support for the AXI PWM Generator. This device is an
FPGA-implemented peripheral used as PWM signal generator and can be
interfaced with AXI4. The register map of this peripheral makes possible
the configuration of the output signal, in terms of period and duty
cycle.

Signed-off-by: Sergiu Cuciurean <sergiu.cuciurean@analog.com>
(cherry picked from commit ff7cf3e87cb5ba0d667c8062cf74e022bd6c229b)