iio: adi-axi-adc: Add of_node to converter structure
The goal of the of_node pointer is to give more flexibility when linking
clients to the AXI ADC. For example, for devices that might have more than
one ADC/port where more than one instance of the axi-adc is probed, we
need to be able to attach to the correct client. Hence, if we just the
of_node of the device that registered with us (and if that device has multiple
clients) we can't really distinguish who is who.

A possible  example of this issue are transceivers with more than one receive
path.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>